+++ src/main/java/com/runapp/achievementservice/service/GoalService.java
@@ -3,6 +3,7 @@ package com.runapp.achievementservice.service;
 import com.runapp.achievementservice.exception.NoEntityFoundException;
 import com.runapp.achievementservice.model.GoalModel;
 import com.runapp.achievementservice.repository.GoalRepository;
+import com.runapp.achievementservice.service.goalUpdater.GoalUpdater;
 import com.runapp.achievementservice.util.supportClasses.GoalMark;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -28,7 +29,11 @@ public class GoalService {
     }
 
     public void deleteGoalById(int id) {
-        goalRepository.deleteById(id);
+        if (goalRepository.existsById(id)) {
+            goalRepository.deleteById(id);
+        } else {
+            throw new NoEntityFoundException("Goal not found with id: " + id);
+        }
     }
 
     public GoalModel updateGoal(GoalModel updatedGoal) {
@@ -37,11 +42,5 @@ public class GoalService {
         } else {
             throw new NoEntityFoundException("Goal not found with id: " + updatedGoal.getId());
         }
-    }
-
-    public GoalModel markGoalCompleted(int goalId) {
-        GoalModel goalModel = goalRepository.findById(goalId)
-                .orElseThrow(() -> new NoEntityFoundException("Goal not found with id: " + goalId));
-        return GoalMark.finishGoal(goalModel);
     }
 }
